+++ pkg/azuremachinepool/mutate_create.go
@@ -5,77 +5,23 @@ import (
 
 	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2020-06-01/compute"
 	"github.com/giantswarm/microerror"
-	"github.com/giantswarm/micrologger"
-	"k8s.io/api/admission/v1beta1"
 	capzexp "sigs.k8s.io/cluster-api-provider-azure/exp/api/v1alpha3"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/azure-admission-controller/internal/patches"
 	"github.com/giantswarm/azure-admission-controller/internal/vmcapabilities"
-	"github.com/giantswarm/azure-admission-controller/pkg/generic"
+	"github.com/giantswarm/azure-admission-controller/pkg/key"
 	"github.com/giantswarm/azure-admission-controller/pkg/mutator"
 )
 
-type CreateMutator struct {
-	ctrlClient client.Client
-	location   string
-	logger     micrologger.Logger
-	vmcaps     *vmcapabilities.VMSKU
-}
-
-type CreateMutatorConfig struct {
-	CtrlClient client.Client
-	Location   string
-	Logger     micrologger.Logger
-	VMcaps     *vmcapabilities.VMSKU
-}
-
-func NewCreateMutator(config CreateMutatorConfig) (*CreateMutator, error) {
-	if config.CtrlClient == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.CtrlClient must not be empty", config)
-	}
-	if config.Location == "" {
-		return nil, microerror.Maskf(invalidConfigError, "%T.Location must not be empty", config)
-	}
-	if config.Logger == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
-	}
-	if config.VMcaps == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.VMcaps must not be empty", config)
-	}
-
-	m := &CreateMutator{
-		ctrlClient: config.CtrlClient,
-		location:   config.Location,
-		logger:     config.Logger,
-		vmcaps:     config.VMcaps,
-	}
-
-	return m, nil
-}
-
-func (m *CreateMutator) Mutate(ctx context.Context, request *v1beta1.AdmissionRequest) ([]mutator.PatchOperation, error) {
+func (h *WebhookHandler) OnCreateMutate(ctx context.Context, object interface{}) ([]mutator.PatchOperation, error) {
 	var result []mutator.PatchOperation
-
-	if request.DryRun != nil && *request.DryRun {
-		m.logger.LogCtx(ctx, "level", "debug", "message", "Dry run is not supported. Request processing stopped.")
-		return result, nil
-	}
-
-	azureMPCR := &capzexp.AzureMachinePool{}
-	if _, _, err := mutator.Deserializer.Decode(request.Object.Raw, nil, azureMPCR); err != nil {
-		return []mutator.PatchOperation{}, microerror.Maskf(parsingFailedError, "unable to parse azureMachinePool CR: %v", err)
-	}
-
-	capi, err := generic.IsCAPIRelease(azureMPCR)
+	azureMPCR, err := key.ToAzureMachinePoolPtr(object)
 	if err != nil {
 		return []mutator.PatchOperation{}, microerror.Mask(err)
 	}
-	if capi {
-		return []mutator.PatchOperation{}, nil
-	}
+	azureMPCROriginal := azureMPCR.DeepCopy()
 
-	patch, err := m.ensureLocation(ctx, azureMPCR)
+	patch, err := h.ensureLocation(ctx, azureMPCR)
 	if err != nil {
 		return []mutator.PatchOperation{}, microerror.Mask(err)
 	}
@@ -83,7 +29,7 @@ func (m *CreateMutator) Mutate(ctx context.Context, request *v1beta1.AdmissionRe
 		result = append(result, *patch)
 	}
 
-	patch, err = m.ensureStorageAccountType(ctx, azureMPCR)
+	patch, err = h.ensureStorageAccountType(ctx, azureMPCR)
 	if err != nil {
 		return []mutator.PatchOperation{}, microerror.Mask(err)
 	}
@@ -91,7 +37,7 @@ func (m *CreateMutator) Mutate(ctx context.Context, request *v1beta1.AdmissionRe
 		result = append(result, *patch)
 	}
 
-	patch, err = m.ensureDataDisks(ctx, azureMPCR)
+	patch, err = h.ensureDataDisks(ctx, azureMPCR)
 	if err != nil {
 		return []mutator.PatchOperation{}, microerror.Mask(err)
 	}
@@ -99,7 +45,7 @@ func (m *CreateMutator) Mutate(ctx context.Context, request *v1beta1.AdmissionRe
 		result = append(result, *patch)
 	}
 
-	patch, err = mutator.EnsureReleaseVersionLabel(ctx, m.ctrlClient, azureMPCR.GetObjectMeta())
+	patch, err = mutator.EnsureReleaseVersionLabel(ctx, h.ctrlClient, azureMPCR.GetObjectMeta())
 	if err != nil {
 		return []mutator.PatchOperation{}, microerror.Mask(err)
 	}
@@ -107,7 +53,7 @@ func (m *CreateMutator) Mutate(ctx context.Context, request *v1beta1.AdmissionRe
 		result = append(result, *patch)
 	}
 
-	patch, err = mutator.CopyAzureOperatorVersionLabelFromAzureClusterCR(ctx, m.ctrlClient, azureMPCR.GetObjectMeta())
+	patch, err = mutator.CopyAzureOperatorVersionLabelFromAzureClusterCR(ctx, h.ctrlClient, azureMPCR.GetObjectMeta())
 	if err != nil {
 		return []mutator.PatchOperation{}, microerror.Mask(err)
 	}
@@ -118,7 +64,7 @@ func (m *CreateMutator) Mutate(ctx context.Context, request *v1beta1.AdmissionRe
 	azureMPCR.Default()
 	{
 		var capiPatches []mutator.PatchOperation
-		capiPatches, err = patches.GenerateFrom(request.Object.Raw, azureMPCR)
+		capiPatches, err = patches.GenerateFromObjectDiff(azureMPCROriginal, azureMPCR)
 		if err != nil {
 			return []mutator.PatchOperation{}, microerror.Mask(err)
 		}
@@ -131,15 +77,7 @@ func (m *CreateMutator) Mutate(ctx context.Context, request *v1beta1.AdmissionRe
 	return result, nil
 }
 
-func (m *CreateMutator) Log(keyVals ...interface{}) {
-	m.logger.Log(keyVals...)
-}
-
-func (m *CreateMutator) Resource() string {
-	return "azuremachinepool"
-}
-
-func (m *CreateMutator) ensureStorageAccountType(ctx context.Context, mpCR *capzexp.AzureMachinePool) (*mutator.PatchOperation, error) {
+func (h *WebhookHandler) ensureStorageAccountType(ctx context.Context, mpCR *capzexp.AzureMachinePool) (*mutator.PatchOperation, error) {
 	if mpCR.Spec.Template.OSDisk.ManagedDisk.StorageAccountType == "" {
 		// We need to set the default value as it is missing.
 
@@ -147,11 +85,11 @@ func (m *CreateMutator) ensureStorageAccountType(ctx context.Context, mpCR *capz
 		if location == "" {
 			// The location was empty and we are adding it using this same mutator.
 			// We assume it will be set to the installation's location.
-			location = m.location
+			location = h.location
 		}
 
 		// Check if the VM has Premium Storage capability.
-		premium, err := m.vmcaps.HasCapability(ctx, location, mpCR.Spec.Template.VMSize, vmcapabilities.CapabilityPremiumIO)
+		premium, err := h.vmcaps.HasCapability(ctx, location, mpCR.Spec.Template.VMSize, vmcapabilities.CapabilityPremiumIO)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -171,7 +109,7 @@ func (m *CreateMutator) ensureStorageAccountType(ctx context.Context, mpCR *capz
 	return nil, nil
 }
 
-func (m *CreateMutator) ensureDataDisks(ctx context.Context, mpCR *capzexp.AzureMachinePool) (*mutator.PatchOperation, error) {
+func (h *WebhookHandler) ensureDataDisks(_ context.Context, mpCR *capzexp.AzureMachinePool) (*mutator.PatchOperation, error) {
 	if len(mpCR.Spec.Template.DataDisks) > 0 {
 		return nil, nil
 	}
@@ -179,10 +117,10 @@ func (m *CreateMutator) ensureDataDisks(ctx context.Context, mpCR *capzexp.Azure
 	return mutator.PatchAdd("/spec/template/dataDisks", desiredDataDisks), nil
 }
 
-func (m *CreateMutator) ensureLocation(ctx context.Context, mpCR *capzexp.AzureMachinePool) (*mutator.PatchOperation, error) {
+func (h *WebhookHandler) ensureLocation(_ context.Context, mpCR *capzexp.AzureMachinePool) (*mutator.PatchOperation, error) {
 	if len(mpCR.Spec.Location) > 0 {
 		return nil, nil
 	}
 
-	return mutator.PatchAdd("/spec/location", m.location), nil
+	return mutator.PatchAdd("/spec/location", h.location), nil
 }