+++ pkg/azuremachinepool/validate_update.go
@@ -4,54 +4,26 @@ import (
 	"context"
 
 	"github.com/giantswarm/microerror"
-	"github.com/giantswarm/micrologger"
-	"k8s.io/api/admission/v1beta1"
 	capzexp "sigs.k8s.io/cluster-api-provider-azure/exp/api/v1alpha3"
 
 	"github.com/giantswarm/azure-admission-controller/internal/vmcapabilities"
 	"github.com/giantswarm/azure-admission-controller/pkg/generic"
-	"github.com/giantswarm/azure-admission-controller/pkg/validator"
+	"github.com/giantswarm/azure-admission-controller/pkg/key"
 )
 
-type UpdateValidator struct {
-	logger micrologger.Logger
-	vmcaps *vmcapabilities.VMSKU
-}
-
-type UpdateValidatorConfig struct {
-	Logger micrologger.Logger
-	VMcaps *vmcapabilities.VMSKU
-}
-
-func NewUpdateValidator(config UpdateValidatorConfig) (*UpdateValidator, error) {
-	if config.Logger == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
+func (h *WebhookHandler) OnUpdateValidate(ctx context.Context, oldObject interface{}, object interface{}) error {
+	azureMPNewCR, err := key.ToAzureMachinePoolPtr(object)
+	if err != nil {
+		return microerror.Mask(err)
 	}
-	if config.VMcaps == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.VMcaps must not be empty", config)
-	}
-
-	admitter := &UpdateValidator{
-		logger: config.Logger,
-		vmcaps: config.VMcaps,
-	}
-
-	return admitter, nil
-}
-
-func (a *UpdateValidator) Validate(ctx context.Context, request *v1beta1.AdmissionRequest) error {
-	azureMPNewCR := &capzexp.AzureMachinePool{}
-	azureMPOldCR := &capzexp.AzureMachinePool{}
-	if _, _, err := validator.Deserializer.Decode(request.Object.Raw, nil, azureMPNewCR); err != nil {
-		return microerror.Maskf(parsingFailedError, "unable to parse azureMachinePool CR: %v", err)
-	}
-	if _, _, err := validator.Deserializer.Decode(request.OldObject.Raw, nil, azureMPOldCR); err != nil {
-		return microerror.Maskf(parsingFailedError, "unable to parse azureMachinePool CR: %v", err)
-	}
-
 	if !azureMPNewCR.GetDeletionTimestamp().IsZero() {
-		a.logger.LogCtx(ctx, "level", "debug", "message", "The object is being deleted so we don't validate it")
+		h.logger.LogCtx(ctx, "level", "debug", "message", "The object is being deleted so we don't validate it")
 		return nil
+	}
+
+	azureMPOldCR, err := key.ToAzureMachinePoolPtr(oldObject)
+	if err != nil {
+		return microerror.Mask(err)
 	}
 
 	capi, err := generic.IsCAPIRelease(azureMPNewCR)
@@ -72,27 +44,27 @@ func (a *UpdateValidator) Validate(ctx context.Context, request *v1beta1.Admissi
 		return microerror.Mask(err)
 	}
 
-	err = checkInstanceTypeIsValid(ctx, a.vmcaps, azureMPNewCR)
+	err = checkInstanceTypeIsValid(ctx, h.vmcaps, azureMPNewCR)
 	if err != nil {
 		return microerror.Mask(err)
 	}
 
-	err = a.checkAcceleratedNetworkingUpdateIsValid(ctx, azureMPOldCR, azureMPNewCR)
+	err = h.checkAcceleratedNetworkingUpdateIsValid(ctx, azureMPOldCR, azureMPNewCR)
 	if err != nil {
 		return microerror.Mask(err)
 	}
 
-	err = a.checkInstanceTypeChangeIsValid(ctx, azureMPOldCR, azureMPNewCR)
+	err = h.checkInstanceTypeChangeIsValid(ctx, azureMPOldCR, azureMPNewCR)
 	if err != nil {
 		return microerror.Mask(err)
 	}
 
-	err = a.checkSpotVMOptionsUnchanged(ctx, azureMPOldCR, azureMPNewCR)
+	err = h.checkSpotVMOptionsUnchanged(ctx, azureMPOldCR, azureMPNewCR)
 	if err != nil {
 		return microerror.Mask(err)
 	}
 
-	err = a.checkStorageAccountTypeUnchanged(ctx, azureMPOldCR, azureMPNewCR)
+	err = h.checkStorageAccountTypeUnchanged(ctx, azureMPOldCR, azureMPNewCR)
 	if err != nil {
 		return microerror.Mask(err)
 	}
@@ -115,7 +87,7 @@ func (a *UpdateValidator) Validate(ctx context.Context, request *v1beta1.Admissi
 	return nil
 }
 
-func (a *UpdateValidator) checkAcceleratedNetworkingUpdateIsValid(ctx context.Context, azureMPOldCR *capzexp.AzureMachinePool, azureMPNewCR *capzexp.AzureMachinePool) error {
+func (h *WebhookHandler) checkAcceleratedNetworkingUpdateIsValid(ctx context.Context, azureMPOldCR *capzexp.AzureMachinePool, azureMPNewCR *capzexp.AzureMachinePool) error {
 	if hasAcceleratedNetworkingPropertyChanged(ctx, azureMPOldCR, azureMPNewCR) {
 		return microerror.Maskf(acceleratedNetworkingWasChangedError, "It is not possible to change the AcceleratedNetworking on an existing node pool")
 	}
@@ -124,7 +96,7 @@ func (a *UpdateValidator) checkAcceleratedNetworkingUpdateIsValid(ctx context.Co
 		return nil
 	}
 
-	err := checkAcceleratedNetworking(ctx, a.vmcaps, azureMPNewCR)
+	err := checkAcceleratedNetworking(ctx, h.vmcaps, azureMPNewCR)
 	if err != nil {
 		return microerror.Mask(err)
 	}
@@ -132,14 +104,14 @@ func (a *UpdateValidator) checkAcceleratedNetworkingUpdateIsValid(ctx context.Co
 	return nil
 }
 
-func (a *UpdateValidator) checkInstanceTypeChangeIsValid(ctx context.Context, azureMPOldCR *capzexp.AzureMachinePool, azureMPNewCR *capzexp.AzureMachinePool) error {
+func (h *WebhookHandler) checkInstanceTypeChangeIsValid(ctx context.Context, azureMPOldCR *capzexp.AzureMachinePool, azureMPNewCR *capzexp.AzureMachinePool) error {
 	// Check if the instance type has changed.
 	if azureMPOldCR.Spec.Template.VMSize != azureMPNewCR.Spec.Template.VMSize {
-		oldPremium, err := a.vmcaps.HasCapability(ctx, azureMPOldCR.Spec.Location, azureMPOldCR.Spec.Template.VMSize, vmcapabilities.CapabilityPremiumIO)
+		oldPremium, err := h.vmcaps.HasCapability(ctx, azureMPOldCR.Spec.Location, azureMPOldCR.Spec.Template.VMSize, vmcapabilities.CapabilityPremiumIO)
 		if err != nil {
 			return microerror.Mask(err)
 		}
-		newPremium, err := a.vmcaps.HasCapability(ctx, azureMPNewCR.Spec.Location, azureMPNewCR.Spec.Template.VMSize, vmcapabilities.CapabilityPremiumIO)
+		newPremium, err := h.vmcaps.HasCapability(ctx, azureMPNewCR.Spec.Location, azureMPNewCR.Spec.Template.VMSize, vmcapabilities.CapabilityPremiumIO)
 		if err != nil {
 			return microerror.Mask(err)
 		}
@@ -154,7 +126,7 @@ func (a *UpdateValidator) checkInstanceTypeChangeIsValid(ctx context.Context, az
 	return nil
 }
 
-func (a *UpdateValidator) checkSpotVMOptionsUnchanged(ctx context.Context, azureMPOldCR *capzexp.AzureMachinePool, azureMPNewCR *capzexp.AzureMachinePool) error {
+func (h *WebhookHandler) checkSpotVMOptionsUnchanged(_ context.Context, azureMPOldCR *capzexp.AzureMachinePool, azureMPNewCR *capzexp.AzureMachinePool) error {
 
 	switch {
 	case (azureMPOldCR.Spec.Template.SpotVMOptions == nil && azureMPNewCR.Spec.Template.SpotVMOptions == nil):
@@ -177,14 +149,10 @@ func (a *UpdateValidator) checkSpotVMOptionsUnchanged(ctx context.Context, azure
 }
 
 // Checks if the storage account type of the osDisk is changed. This is never allowed.
-func (a *UpdateValidator) checkStorageAccountTypeUnchanged(ctx context.Context, azureMPOldCR *capzexp.AzureMachinePool, azureMPNewCR *capzexp.AzureMachinePool) error {
+func (h *WebhookHandler) checkStorageAccountTypeUnchanged(_ context.Context, azureMPOldCR *capzexp.AzureMachinePool, azureMPNewCR *capzexp.AzureMachinePool) error {
 	if azureMPOldCR.Spec.Template.OSDisk.ManagedDisk.StorageAccountType != azureMPNewCR.Spec.Template.OSDisk.ManagedDisk.StorageAccountType {
 		return microerror.Maskf(storageAccountWasChangedError, "Changing the storage account type of the OS disk is not allowed.")
 	}
 
 	return nil
-}
-
-func (a *UpdateValidator) Log(keyVals ...interface{}) {
-	a.logger.Log(keyVals...)
 }