+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cluster webhook handler that replaces mutators and validators.
 - AzureCluster webhook handler that replaces mutators and validators.
 - MachinePool webhook handler that replaces mutators and validators.
+- AzureMachinePool webhook handler that replaces mutators and validators.
 
 ### Changed